+++ rbxbundle/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = "0.5.4"
+__version__ = "0.5.5"
 __author__ = "samuelffer"
 __license__ = "MIT"